HW9_API_Serving/PR1_StreamLit/requirements.txt
streamlit==1.29.0
requests==2.31.0
numpy==1.26.0
pandas==2.1.3
Pillow==10.1.0
pytest==7.4.3
pytest-mock==3.12.0
